[Refactor] Refactor clusters Create
