+++ openstack/cce/v3/clusters/results.go
@@ -3,7 +3,7 @@ package clusters
 import (
 	"encoding/json"
 
-	"github.com/opentelekomcloud/gophertelekomcloud"
+	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 )
 
 type ListCluster struct {
@@ -13,136 +13,6 @@ type ListCluster struct {
 	ApiVersion string `json:"apiVersion"`
 	// all Clusters
 	Clusters []Clusters `json:"items"`
-}
-
-type Clusters struct {
-	// API type, fixed value Cluster
-	Kind string `json:"kind" required:"true"`
-	// API version, fixed value v3
-	ApiVersion string `json:"apiversion" required:"true"`
-	// Metadata of a Cluster
-	Metadata MetaData `json:"metadata" required:"true"`
-	// specifications of a Cluster
-	Spec Spec `json:"spec" required:"true"`
-	// status of a Cluster
-	Status Status `json:"status"`
-}
-
-// Metadata required to create a cluster
-type MetaData struct {
-	// Cluster unique name
-	Name string `json:"name"`
-	// Cluster unique Id
-	Id string `json:"uid"`
-	// Cluster tag, key/value pair format
-	Labels map[string]string `json:"labels,omitempty"`
-	// Cluster annotation, key/value pair format
-	Annotations map[string]string `json:"annotations,omitempty"`
-}
-
-// Specifications to create a cluster
-type Spec struct {
-	// Cluster category: CCE, Turbo
-	Category string `json:"category,omitempty"`
-	// Cluster Type: VirtualMachine, BareMetal, or Windows
-	Type string `json:"type" required:"true"`
-	// Cluster specifications
-	Flavor string `json:"flavor" required:"true"`
-	// Cluster's baseline Kubernetes version. The latest version is recommended
-	Version string `json:"version,omitempty"`
-	// Cluster description
-	Description string `json:"description,omitempty"`
-	// Public IP ID
-	PublicIP string `json:"publicip_id,omitempty"`
-	// Node network parameters
-	HostNetwork HostNetworkSpec `json:"hostNetwork" required:"true"`
-	// Container network parameters
-	ContainerNetwork ContainerNetworkSpec `json:"containerNetwork" required:"true"`
-	// ENI network parameters
-	EniNetwork *EniNetworkSpec `json:"eniNetwork,omitempty"`
-	// Authentication parameters
-	Authentication AuthenticationSpec `json:"authentication,omitempty"`
-	// Charging mode of the cluster, which is 0 (on demand)
-	BillingMode int `json:"billingMode,omitempty"`
-	// Extended parameter for a cluster
-	ExtendParam map[string]string `json:"extendParam,omitempty"`
-	// KubernetesSvcIpRange Service CIDR block or the IP address range which the kubernetes clusterIp must fall within.
-	// This parameter is available only for clusters of v1.11.7 and later.
-	KubernetesSvcIpRange string `json:"kubernetesSvcIpRange,omitempty"`
-	// KubeProxyMode Service forwarding mode. One of `iptables`, `ipvs`
-	KubeProxyMode string `json:"kubeProxyMode,omitempty"`
-	// The system disks and data disks of the master nodes in the cluster are encrypted.
-	EnableMasterVolumeEncryption *bool `json:"enableMasterVolumeEncryption,omitempty"`
-}
-
-// Node network parameters
-type HostNetworkSpec struct {
-	// The ID of the VPC used to create the node
-	VpcId string `json:"vpc" required:"true"`
-	// The ID of the subnet used to create the node
-	SubnetId string `json:"subnet" required:"true"`
-	// The ID of the high speed network used to create bare metal nodes.
-	// This parameter is required when creating a bare metal cluster.
-	HighwaySubnet string `json:"highwaySubnet,omitempty"`
-	// The ID of the Security Group used to create the node
-	SecurityGroupId string `json:"SecurityGroup,omitempty"`
-}
-
-// Container network parameters
-type ContainerNetworkSpec struct {
-	// Container network type: overlay_l2 , underlay_ipvlan or vpc-router
-	Mode string `json:"mode" required:"true"`
-	// Container network segment: 172.16.0.0/16 ~ 172.31.0.0/16. If there is a network segment conflict, it will be automatically reselected.
-	Cidr string `json:"cidr,omitempty"`
-}
-
-type EniNetworkSpec struct {
-	// Eni network subnet id
-	SubnetId string `json:"eniSubnetId" required:"true"`
-	// Eni network cidr
-	Cidr string `json:"eniSubnetCIDR" required:"true"`
-}
-
-// Authentication parameters
-type AuthenticationSpec struct {
-	// Authentication mode: rbac , x509 or authenticating_proxy
-	Mode                string            `json:"mode" required:"true"`
-	AuthenticatingProxy map[string]string `json:"authenticatingProxy" required:"true"`
-}
-
-type Status struct {
-	// The state of the cluster
-	Phase string `json:"phase"`
-	// The ID of the Job that is operating asynchronously in the cluster
-	JobID string `json:"jobID"`
-	// Reasons for the cluster to become current
-	Reason string `json:"reason"`
-	// The status of each component in the cluster
-	Conditions Conditions `json:"conditions"`
-	// Kube-apiserver access address in the cluster
-	Endpoints []Endpoints `json:"-"`
-}
-
-type Conditions struct {
-	// The type of component
-	Type string `json:"type"`
-	// The state of the component
-	Status string `json:"status"`
-	// The reason that the component becomes current
-	Reason string `json:"reason"`
-}
-
-type Endpoints struct {
-	// The address accessed within the user's subnet - OpenTelekomCloud
-	Url string `json:"url"`
-	// Public network access address - OpenTelekomCloud
-	Type string `json:"type"`
-	// Internal network address - OTC
-	Internal string `json:"internal"`
-	// External network address - OTC
-	External string `json:"external"`
-	// Endpoint of the cluster to be accessed through API Gateway - OTC
-	ExternalOTC string `json:"external_otc"`
 }
 
 type Certificate struct {
